Example: fix package install issue

+++ examples/project/mnist/setup.sh
@@ -16,16 +16,20 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-TARGE_DIR=$1
+TARGET_DIR=$1
 EXAMPLE_DIR=$2
 DEMO_DIR=$3
 
 main() {
-    for flowdir in ${TARGE_DIR}/src/flowunit/* ; do
+    for flowdir in ${TARGET_DIR}/src/flowunit/* ; do
         if [ ! -d "${flowdir}" ]; then
             continue;
         fi
 
+        if [ -f "${flowdir}/CMakeLists.txt" ]; then
+            continue;
+        fi
+        
         cp ${EXAMPLE_DIR}/flowunit/python/CMakeLists.txt ${flowdir}/CMakeLists.txt
         if [ $? -ne 0 ]; then
             echo "copy cmake to template failed."
@@ -39,14 +43,14 @@ main() {
         fi
     done
 
-    cp ${EXAMPLE_DIR}/project/base/src/graph/CMakeLists.txt ${TARGE_DIR}/src/graph/CMakeLists.txt
+    cp ${EXAMPLE_DIR}/project/base/src/graph/CMakeLists.txt ${TARGET_DIR}/src/graph/CMakeLists.txt
     if [ $? -ne 0 ]; then
         echo "copy cmake to graph failed."
         return 1
     fi
 
-    mv ${TARGE_DIR}/src/graph/mnist.toml.in ${TARGE_DIR}/src/graph/mnist.toml
-    sed -i "s#@DEMO_MNIST_FLOWUNIT_DIR@#@APPLICATION_PATH@/flowunit#g" ${TARGE_DIR}/src/graph/mnist.toml
+    mv ${TARGET_DIR}/src/graph/mnist.toml.in ${TARGET_DIR}/src/graph/mnist.toml
+    sed -i "s#@DEMO_MNIST_FLOWUNIT_DIR@#@APPLICATION_PATH@/flowunit#g" ${TARGET_DIR}/src/graph/mnist.toml
     if [ $? -ne 0 ]; then
         echo "change graph path failed."
         return 1
